Import files of first web app
This is the first web app:
a "nearest word generator", into which you input a “word”, a “delta” and a “number”. It will generate the "number" of words from the dictionary which are "delta" distance away from the input “word”. Distance is the number of operations needed to transform the input-word into a new word. A single operation is defined as one addition, removal or replacement of a single letter.